Move pluginc classloader to a separate jar
