+++ distribution/build.gradle
@@ -163,6 +163,7 @@ configure(distributions) {
       into 'lib'
       from project(':core').jar
       from project(':core').configurations.runtime
+      from { project(':libs:plugin-classloader').jar }
       // delay add tools using closures, since they have not yet been configured, so no jar task exists yet
       from { project(':distribution:tools:launchers').jar }
       from { project(':distribution:tools:plugin-cli').jar }
