+++ libs/plugin-classloader/build.gradle
@@ -1,0 +1,3 @@
+
+
+apply plugin: 'elasticsearch.build'
